only managers can unassign stage assignment

+++ src/Controller/Application/StageController.php
@@ -74,7 +74,7 @@ class StageController extends AbstractController
             $assignmentService->deleteStageAssignments($stage);
             $assignmentService->addAssignment(new Assignment(), $stage, $assignedUser, $user);
             $status = ['status' => 'ok'];
-        } elseif($assignedUser === null) {
+        } elseif($this->isGranted('ROLE_MANAGER') && $assignedUser === null) {
             $this->deleteAssignment($stage, $assignmentService);
             $status = ['status' => 'ok'];
         }
@@ -119,9 +119,11 @@ class StageController extends AbstractController
             default => Role::USER->string()
         };
 
-        $results[] = ['value' => -1, 'text' => " ____ "];
+        $results = [];
 
         if ($this->isGranted('ROLE_MANAGER')) {
+            $results[] = ['value' => -1, 'text' => " ____ "];
+
             $usersWithAccess = $userService->getUsersWithProjectAccess($project, $neededRole);
         } else {
             // You can only assign yourself if you're not manager
